requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-extensions==4.1
django-soft-delete==1.0.18
djangorestframework==3.16.0
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
